Modified server to include get property.
