+++ airbnb-api/server.js
@@ -22,7 +22,8 @@ const CryptoJS = require('crypto-js');
 /*require controllers*/
 const login = require('./controller/login');
 const register = require('./controller/register');
-const property = require('./controller/property');
+const property = require('./controller/Property/property');
+const propertyList = require('./controller/Property/propertyList');
 
 app.use(express.json());
 app.use(cors());
@@ -39,6 +40,11 @@ app.post('/api/register', (req, res) =>
 /*handle login*/
 app.post('/api/login', (req, res) => login.handleLogin(req, res, db_pool, Joi));
 
+//Get all the information about the property, reviews, rooms, and unavailable dates
+app.get('/api/property/:prid', (req, res) =>
+	property.handleProperty(req, res, db_pool)
+);
+
 //Add property
 app.post('/api/property/add-property', (req, res) =>
 	property.handleAddProperty(req, res, db_pool, Joi)
@@ -46,12 +52,12 @@ app.post('/api/property/add-property', (req, res) =>
 
 //Gets all the list of property according to the type
 app.get('/api/property/property-list/:category', (req, res) =>
-	property.handlePropertyList(req, res, db_pool, Joi)
+	propertyList.handlePropertyList(req, res, db_pool, Joi)
 );
 
 //Gets a number of property according to the type
 app.get('/api/property/property-list/:category/:num', (req, res) =>
-	property.handlePropertyListNum(req, res, db_pool, Joi)
+	propertyList.handlePropertyListNum(req, res, db_pool, Joi)
 );
 
 const port = process.env.PORT || 3000;
